feat: connect the PDF code to the data submission code
