+++ app/club-application/SubmitApplication.jsx
@@ -1,6 +1,7 @@
 import ClubApplicationHeaderSection from "./ClubApplicationHeaderSection";
 import { getCompletionPercentage } from "./requiredData";
 import RegistrationFilled from "./registrationFilled";
+import { PDFCreationComponent, generatePDF } from "./MakePdf";
 
 export default function SubmitApplication(props) {
   const { club } = props;
@@ -64,16 +65,17 @@ export default function SubmitApplication(props) {
       alert("Please fill out all the required fields before submitting your application.");
       return;
     }
+    // get the html for the pdf from getPDF
+    const html = generatePDF();
+
     // send all data to the backend
     // get the club json
-    let clubJSON = club.getJSON
-    console.log(clubJSON);
     fetch("/api/submit", {
       method: "POST",
       headers: {
         "Content-Type": "application/json",
       },
-      body: JSON.stringify(form),
+      body: JSON.stringify({ form: form, html: html }),
     })
       .then((response) => {
         console.log(response);
@@ -130,6 +132,7 @@ export default function SubmitApplication(props) {
           Submit Club Application
         </button>
       </div>
+      <PDFCreationComponent jsonObject={form}/> {/* An invisible component that creates a pdf of the current page */}
     </div>
   );
 }
