+++ app/utils/pdfComponent.js
@@ -1,108 +1,10 @@
-function MyComponent() {
-  // the JSON object that will be used to populate the PDF
-
+function PDFComponent(props) {
   const currentDate = new Date().toLocaleDateString();
   let advisorCounter = 0;
   let memberCounter = 0;
 
-  const form = {
-    clubInformation: {
-      clubName: "MJC Business Club",
-      meetingDaysAndTime: "Every Tuesday and Friday at 9am",
-      meetingLocation: "Both",
-      buildingAndRoomNumber: "Center of Advanced Tech Room 213",
-      zoomLink:
-        "https://us02web.zoom.us/j/86540662672?pwd=TWYxL3VENlM2TmhiYndiKy9QaHZndz09",
-    },
-    clubAdvisors: [
-      {
-        name: "Leslie Vaughn",
-        email: "LeslieVaughn@yosemite.edu",
-        phoneNumber: "(209)-505-3894",
-        title: "Business Professor",
-      },
-    ],
-    clubOfficers: [
-      {
-        role: "Club President",
-        name: "Adrean Cajigas",
-        email: "adrean964837@my.yosemite.edu",
-        wNumber: "964837",
-        phoneNumber: "(209)-505-3894",
-        major: "Computer Science",
-        gradeLevel: "Freshman (1st Year)",
-      },
-      {
-        role: "ICC Representative",
-        name: "Benjamin Schoolland",
-        email: "benjamin849082@my.yosemite.edu",
-        wNumber: "894249",
-        phoneNumber: "(209)-690-69420",
-        major: "Computer Science",
-        gradeLevel: "Senior (4th Year)",
-      },
-      {
-        role: "Club Vice President",
-        placeholder: "This club does not have a Club Vice President",
-      },
-      {
-        role: "Club Secretary",
-        name: "Kirill Kovalenko",
-        email: "kirill123456@my.yosemite.edu",
-        wNumber: "123456",
-        phoneNumber: "(209)-231-5431",
-        major: "Business Administration",
-        gradeLevel: "5+ Years",
-      },
-      {
-        role: "Club Treasurer",
-        placeholder: "This club does not have a Club Treasurer",
-      },
-      {
-        role: "Club Social Media Manager",
-        placeholder: "This club does not have a Club Social Media Manager",
-      },
-    ],
-    clubMembers: [
-      {
-        name: "Adrean Cajigas",
-        email: "adrean.cajigas@gmail.com",
-        wNumber: "0917714",
-      },
-      {
-        name: "Kirill Kovakuku",
-        email: "kuku942145@my.yosemite.edu",
-        wNumber: "696969",
-      },
-      {
-        name: "Beeyn Schoolland",
-        email: "benjiboy@mjc.edu",
-        wNumber: "420420",
-      },
-      {
-        name: "Mason Partridge",
-        email: "masonthegoat@gmail.com",
-        wNumber: "969696",
-      },
-      {
-        name: "Mariah Carey",
-        email: "AllIWantForChristmasIsYou@gmail.com",
-        wNumber: "123456",
-      },
-    ],
-    clubAgreement: [
-      {
-        role: "Club President",
-        signature: "Adrean Cajigas",
-        date: "123",
-      },
-      {
-        role: "Club Advisor",
-        signature: "Brent Wedge",
-        date: "234",
-      },
-    ],
-  };
+  
+  const form = props.form;
   return (
     <div class="pdf">
       <h1>{form.clubInformation.clubName}</h1>
@@ -311,4 +213,4 @@ function MyComponent() {
   );
 }
 
-export default MyComponent;
+export default PDFComponent;
